# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sabah Park Api Documentation",
        "contact": {
            "email": "prastiyo.beka12@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dashboard/summary": {
            "get": {
                "tags": [
                    "DASHBOARD"
                ],
                "summary": "Get Data Member",
                "operationId": "App\\Http\\Controllers\\Admin\\DashboardController::summary",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/auth/admin/login": {
            "post": {
                "tags": [
                    "Admin Credentials"
                ],
                "summary": "Login Admin",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "gate": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/auth/admin/forgot-password": {
            "post": {
                "tags": [
                    "Admin Credentials"
                ],
                "summary": "Forgot Password Admin",
                "operationId": "App\\Http\\Controllers\\AuthController::forgot_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/auth/member/forgot-password": {
            "post": {
                "tags": [
                    "Member Credentials"
                ],
                "summary": "Forgot Password Member",
                "operationId": "App\\Http\\Controllers\\AuthMemberController::forgot_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}